MASSIVE UPDATE TO WORKFLOW
Faculty slot is now resident2 slot. Idea is that
if resident and resident2 grades match, and
generates consensus,  it decreaes load on faculty

For now, a resident ophthalmologist still needs to be
explicitly given allocation of resident and resident2
slots but we fix that in next step after we have
migradted data once.

The Faculty / senior ophthalmology consultants will be
given the slot of Arbitrators to handle non-matches.